Fixed Movement, Added Falling
Each player now uses the same movement script, with customizable (public
variable) control inputs for each.

Added "Pit" tags for areas in which the ball falls off the platform.
Added falling and respawning.

To Do:
-- If the ball falls on the overlap of two "Pit" triggers, it falls
twice as fast.
-- Movement is still possible while falling. Movement should slow down
immensely while falling. This can probably be fixed by changing the
walkspeed variable as the ball is falling.