Added "Expand" Powerup
Added a powerup which causes an increase in size for five seconds. The
powerup disappears upon use. Potential future edits to this powerup are
how large the increase is and whether or not the mass of the object
changes as well.

Next Steps/Necessary Fixes:
-- "Bomb" powerup. Upon pickup, store a bomb for future use. Using the
bomb colors many tiles around the player.
-- Edits to "pit" detection. Sometimes the ball shrinks to a point where
it no longer is hitting the collider so the full fall is not triggered.
Potential solution is to add a specific collider to the ball for falling
which is very small and located at the center of the ball.