Falling Fixes
The complete falling animation/shrinkage now happens on entering the pit
trigger,  not staying in it. This prevents the player from leaving the
trigger and controlling the smaller ball.

This looks great when the player rolls off the edge, but in the cases
where the player approaches the edge very slowly (i.e. when the game is
being tested) it looks awkward.