Began Network Integration
This is a really big update. The game now has the capability, however
buggy it may be, to be played on two systems via a LAN connection.

First and foremost, testing of this can be done by running the project
in two instances, one being a build of the project and the other being
run in the editor. One selects the top button of the UI overlay, the
other selects the next button down.

All of the network-enabling edits have been implemented in their own
scene, denoted MultiplayerLAN.

There have been many steps forward, some steps backward, and some just
allaround changes to the game in this new scene.

First and foremost is color. The floor tiles have been replaced with
cube gameobjects, and can detect the color of the player and change
accordingly. Necessary fixes to this include:
-- Allowing the player to choose their color.
-- Allowing the player to see the other color chosen (currently shows up
as grey).
-- Tallying score (need to edit how player tags are set/detected).

Next, there are general collision/visual problems which must be
addressed:
-- Collisions do not affect the other player.
-- When the other player falls off the edge, the falling effect
(shrinking) does not occur.

Less important fixes include making the Network UI look better and
adding a restart option to the game.

Much of the networking I have learned is due to Tufts' free access to
Lynda courses. Definitely worth watching further to hopefully fix some
of these issues.